Data and Graphs.
This is really the initial commit but it includes a half-baked solution
to taking random data and putting that into a graph. This is not really
well thought out yet since some of it is hard coded and the api is meh.
But, it is at least a start towards what I'd like to do regarding data
viz.

I can hook this up to ruckus and produce a graph of real data related to
the observed metrics. I could also write a reporter for Metrics (by
Coda) to then compare the request data vs the internal reporting which
could be useful.

All that to sayd, not fully thought out, but could be useful.

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>